analyzer/requirements.txt
polars>=0.19.0
pyarrow>=14.0.0
numpy>=1.24.0
pyyaml>=6.0
pytest>=7.0.0